+++ NEWS.md
@@ -11,6 +11,7 @@
 * Remove some message spam from startup
 * Fix mocks failing with `httptest` if used with `with_mock_api()` vs `with_mock_API()`
 * Add `option(googleAuthR.redirect)` to help publish Shiny apps to some platforms (#136)
+* Add support for using gcloud application-default credentials on Jupyter notebooks via `gar_gce_auth_default()` (#147)
 
 # googleAuthR v0.7.0